split devtools stuff
